api/requirements.txt
PyMySQL==1.1.0
certifi==2024.2.2
flask==3.0.3
pony==0.7.17
bcrypt==4.1.3
Flask-Login==0.6.3
PyJWT==2.8.0
python-dotenv==1.0.1
gunicorn==22.0.0
flask-restful==0.3.9
Flask-Testing==0.8.1
flask_cors==3.0.10
pandas==1.5.3
injector==0.18.0
redis==5.0.6
flask-socketio==5.3.6
